+++ CHANGELOG.md
@@ -33,3 +33,8 @@
 ### SQLServer reader
 
 * Add support for `datetime` data type with old SQLServer edition
+
+### HBase11xsql writer
+
+* Add support for truncating table before inserting records
+* Add support for specifing zookeeper's port, if not presents, it uses `2181` as default port